Further improvements
