+++ nxplx-backend/NxPlx.Services.Configuration/NxPlx.Services.Configuration.csproj
@@ -6,7 +6,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+      <PackageReference Include="System.Text.Json" Version="4.7.0" />
     </ItemGroup>
 
 </Project>
