+++ nxplx-frontend/src/NxPlxStore.ts
@@ -2,5 +2,4 @@ interface NxPlxStore {
     isLoggedIn:boolean
     isAdmin:boolean
     build:string
-    localeReady:boolean
 }
