+++ nxplx-frontend/src/routes/home/index.tsx
@@ -6,11 +6,10 @@ import { Component, h } from 'preact';
 import Helmet from 'preact-helmet';
 import {Link, route} from "preact-router";
 import Loading from '../../components/loading';
-import { imageUrl } from "../../Details";
 import http from '../../Http';
-import { OverviewElement } from "../../models";
 import * as style from './style.css';
 import { translate } from "../../localisation";
+import { imageUrl, OverviewElement } from "../../models";
 
 
 interface Props {}
@@ -67,6 +66,7 @@ export default class Home extends Component<Props, State> {
             .then(async response => {
                 if (response.ok) {
                     const overview = await response.json();
+                    console.log(overview);
                     this.setState({ overview: orderBy(overview, ['title'], ['asc']) });
                 }
             })
